Testing Media Query with Search Bar
I'm using media query to display and hide the address locater search bar
depending on the screen size. Unfortunately this makes the application
glitchy. In full screen mode I can see the search bar but it does not
function. In phone screen size the search bar does not display text but
does function. I suspect this has to due with using the same ID element
. How do I get around this?